@@ -2,11 +2,10 @@
   "name": "@my-product-app/frontend-data-access",
   "version": "0.0.1",
   "peerDependencies": {
-    "@my-product-app/frontend-shared": "*",
-    "@angular/common": "^19.2.0",
-    "@angular/core": "^19.2.0",
+    "@angular/core": "~20.1.6",
     "apollo-angular": "^10.0.3",
-    "@apollo/client": "^3.13.8"
+    "rxjs": "~7.8.0",
+    "jest-preset-angular": "~15.0.0"
   },
   "style": "./styles/main.scss"
 }
