@@ -2,8 +2,14 @@
   "name": "@my-product-app/ui",
   "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": "^19.2.0",
-    "@angular/core": "^19.2.0"
+    "@angular/common": "~20.1.6",
+    "@angular/core": "~20.1.6",
+    "@angular/forms": "~20.1.6",
+    "@angular/material": "~20.1.6",
+    "rxjs": "~7.8.0",
+    "@angular/cdk": "~20.1.6",
+    "@angular/router": "~20.1.6",
+    "jest-preset-angular": "~15.0.0"
   },
   "style": "./styles/main.scss"
 }
